icecompr: refactor iceuncompr into a small library
